두 번째 강의 내용 업데이트
