+++ src/RestaurantContainer.test.jsx
@@ -33,7 +33,10 @@ describe('RestaurantContainer', () => {
       address: '서울시 강남구',
       reviews: [
         {
-          id: 1, name: '테스터', description: '맛있어요', score: 1,
+          id: 1,
+          name: '테스터',
+          description: '맛있어요',
+          score: 1,
         },
       ],
     }));
@@ -82,7 +85,7 @@ describe('RestaurantContainer', () => {
           fireEvent.change(getByLabelText(label), { target: { value } });
 
           expect(dispatch).toBeCalledWith({
-            type: 'changeReviewField',
+            type: 'application/changeReviewField',
             payload: { name, value },
           });
         });
